requirements.txt

#Data Analisys
numpy==1.18.4
pandas==1.0.3
scipy==1.4.1
# opencv-python==4.2.0.34
# pip install tensorflow-gpu


# Web Server
redis==3.5.1
mysqlclient==1.4.6
django==3.0.6
django-mysql==3.5.0
django-environ==0.4.5
django-extensions==2.2.9
django-cors-headers==3.2.1
djangorestframework==3.11.0

# Development
# jupyterlab==2.1.2